requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
arts==2024.4.22
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
Flask==2.3.2
Flask-Hashing==1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.4.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
numpy==1.26.4
openai==1.26.0
openai2==1.7.4
protobuf==4.21.12
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pymongo==4.7.2
PyMySQL==1.1.0
pyppeteer==2.0.0
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.25
tornado==6.4
tqdm==4.66.4
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.18.1
